Forced GL to be at least 4.2 now because of a shitty NVIDIA driver issue (see full commit log for more info).
Samplers in GL are bound to an index, but by default, are all bound to zero (why, I will never know).  A simple solution (which I've seen used many places) is to use shader reflection to process all samplers and assign them to the index in the order that they appear.  I have tested this on an Intel driver, and it (surprisingly) works fine, and obviously it worked for others, too.  However, NVIDIA drivers (I'm currently on GTX 650M @ 353.62, and tested on another machine with a different driver) modify the order of samplers, at least when different samplers are used (e.g. sampler2D followed by samplerCube followed by sampler2D may result in the sampler2Ds being grouped and then followed by the samplerCube).  Obviously this means the user writes their shader and the order of the binding is now potentially different without them knowing, which causes sampling of wrong textures and wrong usage of sampler settings... not good.  For now I'm switching to OpenGL 4.2 which allows binding to be done in the shader, and I've disabled the binding in C++, so all GLSL shaders must now bind manually in the shader (which is arguably cleaner anyway).  I may look into it in the future if I provide a reflection interface at all (e.g. process a shader and have it store samplers and their final indices so the user can retrieve them, but it seems more trouble than it is worth).  fuckin' a...