Heavily reorganized project (see full message for details).
- ciri is now a standalone library made up of smaller libraries:
  - ciri is made up of core and gfx, and gfx is further made up of either gfx_gl or gfx_dx, depending on compilation settings
  - ciri will be further split in the future, and refined as necessary; this was just an intermediate cleanup stage
- gfx libs no longer all link; either GL or DX sub-libraries are built and linked depending on the options:
  - this means gl can be used without dx being linked and vice versa
  - this means THE TARGET APPLICATION NO LONGER PICKS THE GRAPHICS API; IT IS SET IN THE LIBRARY UPON BUILD
- ErrorCodes is now an enum class, but I plan to remove all of the ciri::err out into ciri
- graphics devices are created using ciri::createGraphicsDevice() instead of a factory:
  - the function is defined in gfx but has no implementation; gfx_gl or gfx_dx provide a concrete implementation returning their own concrete instances
- new sublime project

There's probably more that I'm missing here, but that is the general idea of it.  In short, ciri is now a library of smaller libraries, and the graphics are linked in to it depending on build settings.

Note that this is indeed a first attempt at cleanup, and will be improved upon and further cleaned and separated (e.g. window, input, image, whatever, I don't know yet).

Hopefully it doesn't introduce any bugs.