Changed ciri's structure once more.  Read full commit for details.
More detail is specified in the structure.txt.  The main changes are that ciri now defines only interfaces for platform-specific code, and that libraries are externally linked to ciri to fulfill concretions of those interfaces, as well as a method of instantiating their according concrete class.  ciri also now uses shared_ptrs, but I'm sure that I'm using them incorrectly (never really played with them that much).  I will figure them out and replace raw pointers elsewhere (e.g. gfx resource creation) with them.  Compiling ciri in visual studio has a few build options (debug/release gl/dx) which automatically build the right configuration in the right order.  That said, there seems to be a bug in visual studio when switching between the graphics libraries, so one needs to rebuild the appropriate library when initially switching to it, which is okay.